chore: move details screen to jumpstart folder
